requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
gevent==1.4.0
gunicorn==20.0.4
locustio==0.13.5
names==0.3.0
nose2==0.9.1
pymongo==3.10.1
pytest==4.5.0
requests==2.22.0
tavern==0.34.0
termcolor==1.1.0
